move client.py to new file
