+++ src/gocert.py
@@ -3,7 +3,11 @@
 
 """Library for interacting with the GoCert application."""
 
+import logging
+
 import requests
+
+logger = logging.getLogger(__name__)
 
 
 class GoCertClientError(Exception):
@@ -13,21 +17,47 @@ class GoCertClientError(Exception):
 class GoCert:
     """Class to interact with GoCert."""
 
-    def __init__(self, url: str, ca_path: str) -> None:
-        """Initialize a client for interacting with GoCert.
+    API_VERSION = "v1"
 
-        Args:
-            url: the endpoint that gocert is listening on e.g https://gocert.com:8000
-            ca_path: the file path that contains the ca cert that gocert uses for https communication
-        """
-        self.url = url
+    def __init__(self, host: str, ca_path: str) -> None:
+        self.host = host
         self.ca_path = ca_path
+
+    def login(self, username: str, password: str) -> str | None:
+        """Login to gocert by sending the username and password and return a Token."""
+        try:
+            req = requests.post(
+                f"https://{self.host}/login",
+                verify=self.ca_path if self.ca_path else None,
+                json={"username": username, "password": password},
+            )
+        except (requests.RequestException, OSError):
+            return
+        if req.status_code != 200:
+            logger.error("couldn't log in: code %s, %s", req.status_code, req.text)
+            return
+        body = req.text
+        return None if body == "" else body
+
+    def token_is_valid(self, token: str) -> bool:
+        """Return if the token is still valid by attempting to connect to an endpoint."""
+        try:
+            req = requests.get(
+                f"https://{self.host}/accounts",
+                verify=self.ca_path if self.ca_path else None,
+                headers={"Authorization": f"Bearer {token}"},
+            )
+        except (requests.RequestException, OSError):
+            return False
+        if req.status_code == 401:
+            return False
+        return True
 
     def is_api_available(self) -> bool:
         """Return if the GoCert server is reachable."""
         try:
             req = requests.get(
-                f"{self.url}/status",
+                f"https://{self.host}/status",
                 verify=self.ca_path if self.ca_path else None,
             )
         except (requests.RequestException, OSError):
@@ -40,7 +70,7 @@ class GoCert:
         """Return if the GoCert server is initialized."""
         try:
             req = requests.get(
-                f"{self.url}/status",
+                f"https://{self.host}/status",
                 verify=self.ca_path if self.ca_path else None,
             )
         except (requests.RequestException, OSError):
@@ -49,3 +79,29 @@ class GoCert:
             return False
         body = req.json()
         return body.get("initialized", False)
+
+    def create_first_user(self, username: str, password: str) -> int | None:
+        """Create the first admin user.
+
+        Args:
+            username: username of the first user
+            password: password for the first user. It must be longer than 7 characters, have at least one lowercase,
+                one uppercase and one number or special character.
+
+        Returns:
+            int | None: the id of the created user, or None if the request failed
+
+        """
+        try:
+            req = requests.post(
+                f"https://{self.host}/api/{self.API_VERSION}/accounts",
+                verify=self.ca_path if self.ca_path else None,
+                json={"username": username, "password": password},
+            )
+        except (requests.RequestException, OSError):
+            return None
+        if req.status_code != 201:
+            logger.warning("couldn't create first user: code %s, %s", req.status_code, req.text)
+            return None
+        id = req.json().get("id")
+        return int(id) if id else None
